Refactor tests, add custom html parsing
