+++ composeApp/src/commonTest/kotlin/com/duchastel/simon/solenne/fakes/FakeAiModelScope.kt
@@ -1,0 +1,6 @@
+package com.duchastel.simon.solenne.fakes
+
+import com.duchastel.simon.solenne.data.ai.AIModelScope
+import com.duchastel.simon.solenne.data.ai.AIModelScope.GeminiModelScope
+
+val FAKE_AI_MODEL_SCOPE: AIModelScope = GeminiModelScope("do-not-rely-on-this-fake-id")
