+++ composeApp/src/commonTest/kotlin/com/duchastel/simon/solenne/fakes/FakeUrlOpener.kt
@@ -1,0 +1,12 @@
+package com.duchastel.simon.solenne.fakes
+
+import com.duchastel.simon.solenne.util.url.UrlOpener
+
+class FakeUrlOpener : UrlOpener {
+    var lastLaunchedUrl: String? = null
+        private set
+
+    override fun launchUrl(url: String) {
+        lastLaunchedUrl = url
+    }
+}
