+++ composeApp/src/commonTest/kotlin/com/duchastel/simon/solenne/tests/screens/modelproviderconfig/AiModelProviderMappingTest.kt
@@ -1,0 +1,17 @@
+package com.duchastel.simon.solenne.tests.screens.modelproviderconfig
+
+import com.duchastel.simon.solenne.data.ai.AIModelProvider
+import com.duchastel.simon.solenne.screens.modelproviderconfig.toUiModel
+import com.duchastel.simon.solenne.screens.modelproviderselector.UiModelProvider
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+class AiModelProviderMappingTest {
+
+    @Test
+    fun `AIModelProvider Gemini maps to UiModelProvider Gemini`() {
+        val aiModelProvider = AIModelProvider.Gemini
+        val uiModelProvider = aiModelProvider.toUiModel()
+        assertEquals(UiModelProvider.Gemini, uiModelProvider)
+    }
+}
