+++ composeApp/src/commonTest/kotlin/com/duchastel/simon/solenne/util/NavigationTestUtils.kt
