--- composeApp/src/commonTest/kotlin/com/duchastel/simon/solenne/util/fakes/FakeAiApiKeyDb.kt
@@ -1,17 +1,0 @@
-package com.duchastel.simon.solenne.util.fakes
-
-import com.duchastel.simon.solenne.db.aiapikey.AIApiKeyDb
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.MutableStateFlow
-
-class FakeAiApiKeyDb(initialGeminiApiKey: String? = null) : AIApiKeyDb {
-
-    private val geminiApiKeyFlow = MutableStateFlow(initialGeminiApiKey)
-
-    override fun getGeminiApiKeyFlow(): Flow<String?> = geminiApiKeyFlow
-
-    override suspend fun saveGeminiApiKey(apiKey: String): String {
-        geminiApiKeyFlow.value = apiKey
-        return apiKey
-    }
-}
