--- composeApp/src/commonTest/kotlin/com/duchastel/simon/solenne/util/fakes/FakeAiModelScope.kt
@@ -1,6 +1,0 @@
-package com.duchastel.simon.solenne.util.fakes
-
-import com.duchastel.simon.solenne.data.ai.AIModelScope
-import com.duchastel.simon.solenne.data.ai.AIModelScope.GeminiModelScope
-
-val FAKE_AI_MODEL_SCOPE: AIModelScope = GeminiModelScope("do-not-rely-on-this-fake-id")
