--- composeApp/src/commonTest/kotlin/com/duchastel/simon/solenne/util/fakes/FakeUrlOpener.kt
@@ -1,12 +1,0 @@
-package com.duchastel.simon.solenne.util.fakes
-
-import com.duchastel.simon.solenne.util.url.UrlOpener
-
-class FakeUrlOpener : UrlOpener {
-    var lastLaunchedUrl: String? = null
-        private set
-
-    override fun launchUrl(url: String) {
-        lastLaunchedUrl = url
-    }
-}
