+++ gradle/libs.versions.toml
@@ -11,6 +11,7 @@ core = "1.6.1"
 junit = "1.2.1"
 kotlin = "2.1.20"
 coroutines = "1.10.2"
+ksoupHtml = "0.5.0"
 markdown = "0.7.3"
 mcpKotlin = "0.5.0"
 kotlinx-coroutines = "1.10.2"
@@ -26,6 +27,7 @@ multiplatform-settings = "1.3.0"
 androidx-test-core = { module = "androidx.test:core", version.ref = "core" }
 androidx-junit = { module = "androidx.test.ext:junit", version.ref = "junit" }
 circuit-test = { module = "com.slack.circuit:circuit-test", version.ref = "circuit" }
+ksoup-html = { module = "com.mohamedrejeb.ksoup:ksoup-html", version.ref = "ksoupHtml" }
 markdown = { module = "org.jetbrains:markdown", version.ref = "markdown" }
 mcp-kotlin = { module = "com.duchastel.simon.modelcontextprotocol.kotlin.sdk:kotlin-sdk", version.ref = "mcpKotlin" }
 kotlin-test = { module = "org.jetbrains.kotlin:kotlin-test", version.ref = "kotlin" }
